+++ src/Components/Footer/FooterIconData.js
@@ -1,0 +1,29 @@
+const footerIconData = [
+  {
+    id: 1,
+    icon: 'https://res.kurly.com/pc/ico/1810/ico_instagram.png',
+    alt: '초컬리 인스타그램 바로가기',
+  },
+  {
+    id: 2,
+    icon: 'https://res.kurly.com/pc/ico/1810/ico_fb.png',
+    alt: '초컬리 페이스북 바로가기',
+  },
+  {
+    id: 3,
+    icon: 'https://res.kurly.com/pc/ico/1810/ico_blog.png',
+    alt: '초컬리 네이버블로그 바로가기',
+  },
+  {
+    id: 4,
+    icon: 'https://res.kurly.com/pc/ico/1810/ico_naverpost.png',
+    alt: '초컬리 유튜브 바로가기',
+  },
+  {
+    id: 5,
+    icon: 'https://res.kurly.com/pc/ico/1810/ico_youtube.png',
+    alt: '초컬리 유튜브 바로가기',
+  },
+];
+
+export default footerIconData;
