+++ src/Pages/ProductDetails/ProductDetails.js
@@ -1,8 +1,12 @@
 import React, { Component } from 'react';
-
+import Footer from '../../Components/Footer/Footer';
 class ProductDetails extends Component {
   render() {
-    return <div></div>;
+    return (
+      <div>
+        <Footer />
+      </div>
+    );
   }
 }
